+++ tests/test_diff.py
@@ -1401,3 +1401,15 @@ def test_show_json_format(data_archive_readonly, cli_runner):
         assert r.exit_code == 0, r
         # output is compact, no indentation
         assert '"sno.diff/v1+hexwkb": {"' in r.stdout
+
+
+def test_show_shallow_clone(data_archive_readonly, cli_runner, tmp_path, chdir):
+    # just checking you can 'show' the first commit of a shallow clone
+    with data_archive_readonly("points") as original_path:
+        clone_path = tmp_path / "shallow-clone"
+        r = cli_runner.invoke(["clone", "--depth=1", original_path, clone_path])
+        assert r.exit_code == 0, r
+
+        with chdir(clone_path):
+            r = cli_runner.invoke(["show"])
+            assert r.exit_code == 0, r